# backend/internal/postagem/postagemsvc/model.go
package postagemsvc

import (
	"errors"
)

type Postagem struct {
	Titulo       string `json:"titulo"`
	Descricao    string `json:"descricao"`
	ImagemBase64 string `json:"imagem_base64"`
	UserID       string `json:"user_id"`
	Categoria    string `json:"categoria"`
}

func (p *Postagem) Validate() error {
	if p.Titulo == "" {
		return errors.New("o título é obrigatório")
	}

	if p.Descricao == "" {
		return errors.New("a descrição é obrigatória")
	}

	if p.Categoria == "" {
		return errors.New("a categoria é obrigatória")
	}

	return nil
}